fixing JET issues (#261)

+++ src/ProtocolZoo/cutoff.jl
@@ -72,7 +72,7 @@ end
         if isnothing(prot.period)
             @yield onchange_tag(reg)
         else
-            @yield timeout(prot.sim, prot.period)
+            @yield timeout(prot.sim, prot.period::Float64)
         end
     end
 end
